Refactor: Move amount and price related methods to new validation classes
