+++ core/src/test/java/bisq/core/trade/validation/DepositTxValidationTest.java
@@ -21,7 +21,6 @@ import bisq.core.btc.model.RawTransactionInput;
 import bisq.core.btc.wallet.BtcWalletService;
 import bisq.core.btc.wallet.TradeWalletService;
 import bisq.core.dao.burningman.DelayedPayoutTxReceiverService;
-import bisq.core.exceptions.TradePriceOutOfToleranceException;
 import bisq.core.offer.Offer;
 import bisq.core.provider.fee.FeeService;
 import bisq.core.provider.price.PriceFeedService;
@@ -64,12 +63,12 @@ import java.util.List;
 import org.junit.jupiter.api.Test;
 
 import static bisq.core.trade.validation.DepositTxValidation.checkDepositTxMatchesIgnoringWitnessesAndScriptSigs;
+import static bisq.core.trade.validation.TradeAmountValidationTest.OFFER_MIN_AMOUNT;
 import static bisq.core.trade.validation.TradeValidationTestUtils.*;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertSame;
 import static org.junit.jupiter.api.Assertions.assertThrows;
-import static org.mockito.Mockito.doThrow;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
@@ -77,76 +76,10 @@ import static org.mockito.Mockito.when;
 public class DepositTxValidationTest {
     static final String SELLER_ADDRESS = SegwitAddress.fromKey(PARAMS, new ECKey()).toString();
     static final String BUYER_ADDRESS = SegwitAddress.fromKey(PARAMS, new ECKey()).toString();
-    static final Coin OFFER_MIN_AMOUNT = Coin.valueOf(1_000);
-    static final Coin OFFER_MAX_AMOUNT = Coin.valueOf(5_000);
+
     private static final String MAKER_ROLE = "Maker";
     private static final String TAKER_ROLE = "Taker";
 
-    /* --------------------------------------------------------------------- */
-    // TradeAmount
-    /* --------------------------------------------------------------------- */
-
-    @Test
-    void checkTradeAmountAcceptsOfferBoundsAndValuesBetweenThem() {
-        Coin tradeAmount = Coin.valueOf(3_000);
-
-        assertSame(OFFER_MIN_AMOUNT,
-                DepositTxValidation.checkTradeAmount(OFFER_MIN_AMOUNT, OFFER_MIN_AMOUNT, OFFER_MAX_AMOUNT));
-        assertSame(tradeAmount,
-                DepositTxValidation.checkTradeAmount(tradeAmount, OFFER_MIN_AMOUNT, OFFER_MAX_AMOUNT));
-        assertSame(OFFER_MAX_AMOUNT,
-                DepositTxValidation.checkTradeAmount(OFFER_MAX_AMOUNT, OFFER_MIN_AMOUNT, OFFER_MAX_AMOUNT));
-    }
-
-    @Test
-    void checkTradeAmountRejectsAmountsBelowOfferMinimum() {
-        IllegalArgumentException exception = assertThrows(IllegalArgumentException.class,
-                () -> DepositTxValidation.checkTradeAmount(Coin.valueOf(999), OFFER_MIN_AMOUNT, OFFER_MAX_AMOUNT));
-
-        assertEquals("Trade amount must not be less than minimum offer amount. " +
-                        "tradeAmount=0.00000999 BTC, offerMinAmount=0.00001 BTC",
-                exception.getMessage());
-    }
-
-    @Test
-    void checkTradeAmountRejectsAmountsAboveOfferMaximum() {
-        IllegalArgumentException exception = assertThrows(IllegalArgumentException.class,
-                () -> DepositTxValidation.checkTradeAmount(Coin.valueOf(5_001), OFFER_MIN_AMOUNT, OFFER_MAX_AMOUNT));
-
-        assertEquals("Trade amount must not be higher than maximum offer amount. " +
-                        "tradeAmount=0.00005001 BTC, offerMaxAmount=0.00005 BTC",
-                exception.getMessage());
-    }
-
-
-    /* --------------------------------------------------------------------- */
-    // TradePrice
-    /* --------------------------------------------------------------------- */
-
-    @Test
-    void checkTakersTradePriceAcceptsVerifiedPrice() throws Exception {
-        long takersTradePrice = 50_000_000L;
-        Offer offer = mock(Offer.class);
-        when(offer.isUseMarketBasedPrice()).thenReturn(true);
-
-        assertEquals(takersTradePrice, DepositTxValidation.checkTakersTradePrice(takersTradePrice,
-                mock(PriceFeedService.class),
-                offer));
-        verify(offer).verifyTakersTradePrice(takersTradePrice);
-    }
-
-    @Test
-    void checkTakersTradePriceWrapsOfferPriceValidationFailure() throws Exception {
-        long takersTradePrice = 50_000_000L;
-        Offer offer = mock(Offer.class);
-        doThrow(new TradePriceOutOfToleranceException("price outside tolerance"))
-                .when(offer)
-                .verifyTakersTradePrice(takersTradePrice);
-
-        assertThrows(RuntimeException.class, () -> DepositTxValidation.checkTakersTradePrice(takersTradePrice,
-                mock(PriceFeedService.class),
-                offer));
-    }
 
 
     /* --------------------------------------------------------------------- */
